Initial attempt at paging through all posts for a topic
A better approach would be some sort of pagination, to handle
long threads, but I haven't thought through this: should it

 - indicate there's more posts available so that if you
   go to the "next" record it downloads the next chunk
 - set up background jobs to load all the data and then these
   can be queried at each "action" to see if they can be added

I think the former is preferable from a user's perspective (i.e.
demand the next set) otherwise how do we indicate to the user
that we are waiting for data to download?