Scroll a single post with shift + up/down
I had hoped it would also work for the home/end and pgup/down,
but it's not working.

It would also be nice to identify visually those posts for which
it's worth scrolling - and then this could identify when we want
to switch between scrolling within the post and between posts.