Initial support for a category page
Pressing c will show the category list and you cna enter into
a category to see the posts. You can't view the posts. The logic
is also messy as we do too much, and should send around the
category map, allowing for missing categories (in case one
gets added after the program starts running).

The pages now tend to have a "banner bar" - that is, a centered
text label - to make it look "more professional".