Rework downloading extra messages
When there are > 20 messages in a post we start a reader thread
that downloads the posts and sends them as an event to be
processed. This is easier to reaon about than the previous
version, which ended up forcing the TUI to wait until all the
messages had been downloaded [*] even when that was not the
intended case.

[*] the code could have been re-written to avoid this but it's
    not worth it when this is a more-sensible approach